+++ uv.gyp
@@ -274,19 +274,43 @@
           },
         }],
         [ 'OS=="aix"', {
-          'sources': [ 'src/unix/aix.c' ],
+          'variables': {
+            'os_name': '<!(uname -s)',
+          },
+          'sources': [
+            'src/unix/aix-common.c',
+          ],
           'defines': [
             '_ALL_SOURCE',
             '_XOPEN_SOURCE=500',
             '_LINUX_SOURCE_COMPAT',
             '_THREAD_SAFE',
-            'HAVE_SYS_AHAFS_EVPRODS_H',
           ],
-          'link_settings': {
-            'libraries': [
-              '-lperfstat',
-            ],
-          },
+          'conditions': [
+            [ '"<(os_name)"=="OS400"', {
+              'sources': [
+                'src/unix/ibmi.c',
+                'src/unix/posix-poll.c',
+                'src/unix/no-fsevents.c',
+                'src/unix/no-proctitle.c',
+              ],
+              'defines': [
+                '_PASE=1'
+              ],
+            }, {
+              'sources': [
+                'src/unix/aix.c'
+              ],
+              'defines': [
+                'HAVE_SYS_AHAFS_EVPRODS_H'
+              ],
+              'link_settings': {
+                'libraries': [
+                  '-lperfstat',
+                ],
+              },
+            }],
+          ]
         }],
         [ 'OS=="freebsd" or OS=="dragonflybsd"', {
           'sources': [ 'src/unix/freebsd.c' ],
